requirements.txt
flask==2.3.2
flask-cors==3.0.10
chromadb==0.3.22
sentence-transformers==2.2.2
google-generativeai==0.2.3
azure-storage-blob==12.16.0
